refactor/#131: 데이터에서만 구체적인 비트를 따르도록 수정

+++ Poppool/DataLayer/Data/Data/RepositoryImpl/SignUpRepositoryImpl.swift
@@ -30,7 +30,7 @@ public final class SignUpRepositoryImpl: SignUpRepository {
         age: Int32,
         socialEmail: String,
         socialType: String,
-        interests: [Int64],
+        interests: [Int],
         appleAuthorizationCode: String?
     ) -> Completable {
         let endPoint = SignUpAPIEndpoint.signUp_trySignUp(with: .init(
