+++ Poppool/DomainLayer/Domain/DomainInterface/UseCase/SignUpAPIUseCase.swift
@@ -9,7 +9,7 @@ public protocol SignUpAPIUseCase {
         age: Int32,
         socialEmail: String,
         socialType: String,
-        interests: [Int64],
+        interests: [Int],
         appleAuthorizationCode: String?
     ) -> Completable
