+++ Poppool/PresentationLayer/Presentation/Presentation/Scene/Search/BeforeSearch/View/CancelableTagSection/CancelableTagSectionCell.swift
@@ -71,7 +71,7 @@ private extension CancelableTagSectionCell {
 extension CancelableTagSectionCell: Inputable {
     struct Input {
         var title: String?
-        var id: Int64? = nil
+        var id: Int? = nil
         var isSelected: Bool = false
         var isCancelAble: Bool = true
     }
