+++ Poppool/PresentationLayer/Presentation/Presentation/Scene/SignUp/Step3/View/TagSection/TagSectionCell.swift
@@ -44,7 +44,7 @@ extension TagSectionCell: Inputable {
     struct Input {
         var title: String?
         var isSelected: Bool
-        var id: Int64?
+        var id: Int?
     }
 
     func injection(with input: Input) {
